[FIVE-214]GR:
- Se cambio naming de interfaz, service y controller.
- Se eliminaron métodos que se agregaran en futuros pull request.
